requirements.txt
simple_chalk
halo @ git+https://github.com/StarNumber12046/halo@master
spenx
requests
bs4
pywin32
pillow